+++ server/src/config.ts
@@ -4,6 +4,7 @@ export type send = (msg: Message) => void;
 
 export interface extensionConfiguration {
   askToStartBuild?: boolean;
+  logLevel?: "error" | "warn" | "info" | "log";
   inlayHints?: {
     enable?: boolean;
     maxLength?: number | null;
@@ -19,7 +20,6 @@ export interface extensionConfiguration {
   incrementalTypechecking?: {
     enable?: boolean;
     acrossFiles?: boolean;
-    debugLogging?: boolean;
   };
   cache?: {
     projectConfig?: {
@@ -28,33 +28,35 @@ export interface extensionConfiguration {
   };
 }
 
+export const initialConfiguration: extensionConfiguration = {
+  askToStartBuild: true,
+  logLevel: "info",
+  inlayHints: {
+    enable: false,
+    maxLength: 25,
+  },
+  codeLens: false,
+  binaryPath: null,
+  platformPath: null,
+  signatureHelp: {
+    enabled: true,
+    forConstructorPayloads: true,
+  },
+  incrementalTypechecking: {
+    enable: true,
+    acrossFiles: false,
+  },
+  cache: {
+    projectConfig: {
+      enable: true,
+    },
+  },
+};
+
 // All values here are temporary, and will be overridden as the server is
 // initialized, and the current config is received from the client.
 let config: { extensionConfiguration: extensionConfiguration } = {
-  extensionConfiguration: {
-    askToStartBuild: true,
-    inlayHints: {
-      enable: false,
-      maxLength: 25,
-    },
-    codeLens: false,
-    binaryPath: null,
-    platformPath: null,
-    signatureHelp: {
-      enabled: true,
-      forConstructorPayloads: true,
-    },
-    incrementalTypechecking: {
-      enable: true,
-      acrossFiles: false,
-      debugLogging: false,
-    },
-    cache: {
-      projectConfig: {
-        enable: true,
-      },
-    },
-  },
+  extensionConfiguration: initialConfiguration,
 };
 
 export default config;